+++ src/steps/compute/index.test.ts
@@ -1417,6 +1417,7 @@ describe('#fetchComputeBackendBuckets', () => {
           retentionPeriod: { type: 'string' },
           retentionDate: { type: 'string' },
           public: { type: 'boolean' },
+          isSubjectToObjectAcls: { type: 'boolean' },
           classification: { const: null },
           etag: { type: 'string' },
         },