+++ src/steps/logging/index.test.ts
@@ -70,6 +70,7 @@ describe('#fetchProjectSinks', () => {
           retentionPeriod: { type: 'string' },
           retentionDate: { type: 'string' },
           public: { type: 'boolean' },
+          isSubjectToObjectAcls: { type: 'boolean' },
           classification: { const: null },
           etag: { type: 'string' },
           webLink: { type: 'string' },